messaging_app/requirements.txt
asgiref==3.9.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
PyJWT>=2.8.0
djangorestframework-nested>=0.0.1
